requirements.txt
rendercv[full]>=2.0
bibtexparser>=1.4,<2
pyyaml>=6.0
